+++ app/src/main/AndroidManifest.xml
@@ -10,13 +10,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.TD1Collector"
+        android:theme="@style/Theme.FilamentLibrary"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.TD1Collector">
+            android:theme="@style/Theme.FilamentLibrary">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -26,7 +26,7 @@
         <receiver android:name=".util.impl.PermissionReceiver"
             android:exported="true">
             <intent-filter>
-                <action android:name="com.pluralmakes.td1_collector.USB_PERMISSION" />
+                <action android:name="com.pluralmakes.filamentlibrary.USB_PERMISSION" />
             </intent-filter>
         </receiver>
     </application>
