+++ app/src/main/java/com/pluralmakes/filamentlibrary/util/extensions/ColorExtensions.kt
@@ -1,7 +1,6 @@
-package com.pluralmakes.td1_collector.util.extensions
+package com.pluralmakes.filamentlibrary.util.extensions
 
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.toArgb
 
 fun Color.isColorLight(): Boolean {
     val luminance = 0.299 * red + 0.587 * green + 0.114 * blue
