+++ app/src/main/java/com/pluralmakes/filamentlibrary/util/extensions/NumberExtensions.kt
@@ -1,4 +1,4 @@
-package com.pluralmakes.td1_collector.util.extensions
+package com.pluralmakes.filamentlibrary.util.extensions
 
 fun Number.toTDString(): String {
     return "%.1f".format(this)
