+++ app/src/main/java/com/pluralmakes/filamentlibrary/viewModel/CollectorViewModel.kt
@@ -1,14 +1,13 @@
-package com.pluralmakes.td1_collector.viewModel
+package com.pluralmakes.filamentlibrary.viewModel
 
 import android.content.Context
-import android.content.Intent
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
-import com.pluralmakes.td1_collector.model.Filament
-import com.pluralmakes.td1_collector.util.ConnectionStatus
-import com.pluralmakes.td1_collector.util.ConnectionStatus.*
-import com.pluralmakes.td1_collector.util.TD1Communicator
+import com.pluralmakes.filamentlibrary.model.Filament
+import com.pluralmakes.filamentlibrary.util.ConnectionStatus
+import com.pluralmakes.filamentlibrary.util.ConnectionStatus.*
+import com.pluralmakes.filamentlibrary.util.TD1Communicator
 
 class CollectorViewModel(
     val connectionStatus: MutableState<ConnectionStatus> = mutableStateOf(NONE),
