+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">TD-1 Collector</string>
+    <string name="app_name">Filament Library</string>
 </resources>
